renames data pipeline again
trying to be more modular - this pipeline is dedicated to fetching the
data. what we want to do in the next pipeline is to make flattened csv
data for each json. Then we can combine all our normalized csvs into a
csv. That will be the input for the last pipeline