bronze data pipeline
this is necessary because I do not know how to create a function that
takes all of the keyword arguments dynamically. For example - if I call
an argument (**kwargs), the kedro framework expects me to pass an input
called "kwargs".

So in order to ensure that all documents have been downloaded and
transcoded, I have made a separate pipeline that is executed after the
data_fetching and data_processing_intermediate pipelines.

Having all the data together in the "lightly massaged" columns fits the
definition of bronze data, I believe.