filters columns for bronze dataset
this filtering is done by creating a column analysis dataset and then
using the data from the percentage_of_max to make a cut on the column.
The current cut is made at < %50 of the countries having data in the
column

Also included is the data-exploration ipythonb notebook. This notebook
helps show how I came to some of the node implementations.