adds name_dataset to hooks and data_fetching
This commit is an experiment in adding a MemoryDataSet. The existing
country name is nested under
government_country_name_conventional_short_form. The commented code
could add the country_name as a top level key. Hooks has been modified
to create the dataset, and pipeline/nodes for data_fetching has been
modified to allow input of the name as a dataset