adds DTypedSQLDataSet to meet catalog expectations
The DTypedSQLDataSet allows the catalog to specify datatypes correctly
so that, when specified, the dtypes are used in the table definition.
The key difference, is that importlib and getaddr are used to get
references to the actual class and not the string name. the
_modify_save_args_dtype method modifies _save_args so that class
references are used for each column.